Assignment 4/src/ManagementCompanyTest.java
import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class ManagementCompanyTest {

	Property sampleProperty;
	Property sampleProperty1;
	Property sampleProperty2;
	ManagementCompany mangementCo ; 
	
	@BeforeEach
	public void setUp() throws Exception {
		mangementCo= new ManagementCompany("Railey", "555555555",6);
	}

	@AfterEach
	public void tearDown() throws Exception {
		mangementCo=null;
	}

	@Test
	public void testAddProperty() {
		sampleProperty = new Property ("Sunsational", "Beckman", 2613, "BillyBob Wilson",2,5,2,2);
		sampleProperty1 = new Property("Sunsational1", "Beckman1", 2613, "BillyBob Wilson",1,4,2,2);
		sampleProperty2 = new Property("Sunsational1", "Beckman1", 2613, "BillyBob Wilson",20,50,1,1);
		assertEquals(mangementCo.addProperty(sampleProperty),0,0);	//property has been successfully added to index 0
		assertEquals(mangementCo.addProperty(sampleProperty1),-4,0);
		assertEquals(-3,mangementCo.addProperty(sampleProperty2),0);
	}
	
	@Test
	public void testGetPropertiesCount() {
		sampleProperty = new Property ("Sunsational", "Beckman", 2613, "BillyBob Wilson",2,5,2,2);		 
		assertEquals(mangementCo.addProperty(sampleProperty),0,0);	
		assertEquals(mangementCo.getPropertiesCount(), 1);
	}

	@Test
	public void testToString() {
		sampleProperty = new Property ("Sunsational", "Beckman", 2613.0, "BillyBob Wilson",2,5,2,2);
		assertEquals(mangementCo.addProperty(sampleProperty), 0);	//property has been successfully added to index 0
		String expectedString = "List of the properties for Railey, taxID: 555555555\n"
				+ "______________________________________________________\n"
				+ "Sunsational,Beckman,BillyBob Wilson,2613.0\n"
				+ "______________________________________________________\n"
				+"\n"
				+ " total management Fee: 156.78";
		System.out.println(expectedString);
		System.out.println(mangementCo.toString());
		assertEquals(expectedString, mangementCo.toString());
				
	}

}
